chore(package): Stop requiring tslib
With the current es target (es2018), it seems like the current codebase does not require any typescript helpers. This was found when investigating why a tree-shaked es version of the package is much smaller than a cjs version. One reason was that tslib could be completely tree-shaken away since none of the methods from it was used. By changing "importTslib" to false, typescript will insert helpers as needed, but for now, we don't need any and thus does not gain any more package size. If we use any feature that would require ts helpers, it'd be good to again require tslib and add `importTslib` to the tsconfig again.