(Read buffer directly rather than make a copy of it):
Instead of creating a copy of the Buffer, we now read directly from the Buffer. This should be ok since we check the length of the buffer for every operation. This makes for more expected behaviour, and if you want to replicate old behaviour, you can still provide Walking Buffer with a copy using `Buffer.from()`.  We could not come up with any way of making a buffer shorter, meaning there should be no case where the cursor ends up being outside the buffer because the buffer shrank. A side bonus is also less memory overhead for `WalkingBuffer`.

BREAKING CHANGE: WalkingBuffer now reads directly from the buffer provided and does not make a copy of it anymore. To replicate old behaviour, use `new WalkingBuffer({ buffer: Buffer.from(buffer) })`.